requirements.txt
confluent-kafka
azure-eventhub